+++ src/graph/vertex.cpp
@@ -31,7 +31,7 @@ Vertex::Vertex(GraphArtifact *parent)
 
 }
 
-Vertex::Vertex(unsigned int id, GraphArtifact *parent)
+Vertex::Vertex(unsigned long long id, GraphArtifact *parent)
     : GraphArtifact(id, parent)
 {
